Fix some auth functions not using IAuthFunctions
Various auth functions are purely wrappers around other auth functions.
Unfortunately this means when overriding one of auth functions (a parent) that is used by others (children),
the children will still call the original version, not the overridden version.

This commit fixes this for most cases.

Auth functions that are purely a wrapper around a parent auth
function are simplified by using a wrapper function.